// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  availability: typeof availability;
  bookings: typeof bookings;
  categories: typeof categories;
  consultantProfiles: typeof consultantProfiles;
  consultantServices: typeof consultantServices;
  filteredServices: typeof filteredServices;
  http: typeof http;
  notifications: typeof notifications;
  reviews: typeof reviews;
  stream: typeof stream;
  user: typeof user;
}>;
declare const fullApiWithMounts: typeof fullApi;

export declare const api: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "internal">
>;